feature(service): add QueueNumber to QueueProxy.
Allows each client to manage their own queue in the server. Without it, it would be impossible to know which queue belonged to which client. Before, each client had an id assigned locally by the server. However, this would be lost after a connection got closed (i.e. a known client would be assigned a new id, losing management of its existing queue).